Decouple Operation Reporting
